MAINT: Avoid code duplicates for test file directories
Includes some additional refactoring of test utilities.
