+++ tests/bench.py
@@ -5,7 +5,6 @@ The results are on https://py-pdf.github.io/pypdf/dev/bench/
 Please keep in mind that the variance is high.
 """
 from io import BytesIO
-from pathlib import Path
 from tempfile import NamedTemporaryFile
 
 import pytest
@@ -14,12 +13,7 @@ import pypdf
 from pypdf import PageObject, PdfReader, PdfWriter, Transformation
 from pypdf.generic import Destination, read_string_from_stream
 
-from . import get_data_from_url
-
-TESTS_ROOT = Path(__file__).parent.resolve()
-PROJECT_ROOT = TESTS_ROOT.parent
-RESOURCE_ROOT = PROJECT_ROOT / "resources"
-SAMPLE_ROOT = PROJECT_ROOT / "sample-files"
+from . import RESOURCE_ROOT, SAMPLE_ROOT, get_data_from_url
 
 
 def page_ops(pdf_path, password):
