+++ tests/test_cmap.py
@@ -1,6 +1,5 @@
 """Test the pypdf_cmap module."""
 from io import BytesIO
-from pathlib import Path
 
 import pytest
 
@@ -10,11 +9,7 @@ from pypdf._codecs import charset_encoding
 from pypdf._font import Font
 from pypdf.generic import ArrayObject, DictionaryObject, IndirectObject, NameObject, NullObject
 
-from . import get_data_from_url
-
-TESTS_ROOT = Path(__file__).parent.resolve()
-PROJECT_ROOT = TESTS_ROOT.parent
-RESOURCE_ROOT = PROJECT_ROOT / "resources"
+from . import RESOURCE_ROOT, get_data_from_url
 
 
 @pytest.mark.enable_socket
